+++ src/Http/routes.php
@@ -1,5 +1,7 @@
 <?php
 
+use App\Mail\Csdcsdc;
+
 Route::get('/', function(){
 
 	return redirect()->route('mailableList');