+++ changelog.md
@@ -9,3 +9,4 @@ All notable changes to `laravel-mail-editor` will be documented in this file.
 - Fix error `Call to a member function sortBy() on null` that happens when mailables directory not created yet.
 - Replace jquery $.ajax with axios library, also adding notie confirmation for save-template action
 - Fix javascript error `plaintextEditor is not defined` when save mailable view template.
+- Fix mailable with multiple constructor args types error ([issue](https://github.com/Qoraiche/laravel-mail-editor/issues/1)).